+++ tests/linen/toplevel_test.py
@@ -19,7 +19,7 @@ from jax import random
 from jax.nn import initializers
 import jax.numpy as jnp
 
-import numpy as onp
+import numpy as np
 from typing import Any, Tuple
 
 from flax import linen as nn